+++ Mailer/app/Models/Generators/MediaBriefingGenerator.php
@@ -7,7 +7,7 @@ use Nette\Application\UI\Form;
 use Nette\Utils\ArrayHash;
 use Remp\Mailer\Components\GeneratorWidgets\Widgets\MediaBriefingWidget\MediaBriefingWidget;
 use Remp\MailerModule\Models\ContentGenerator\Engine\EngineFactory;
-use Remp\MailerModule\Models\Generators\ArticleLocker;
+use Remp\MailerModule\Models\Generators\HtmlArticleLocker;
 use Remp\MailerModule\Models\Generators\EmbedParser;
 use Remp\MailerModule\Models\Generators\IGenerator;
 use Remp\MailerModule\Models\Generators\PreprocessException;
@@ -39,11 +39,11 @@ class MediaBriefingGenerator implements IGenerator
 
     public function __construct(
         SourceTemplatesRepository $mailSourceTemplateRepository,
-        WordpressHelpers $helpers,
-        ContentInterface $content,
-        EmbedParser $embedParser,
-        ArticleLocker $articleLocker,
-        EngineFactory  $engineFactory
+        WordpressHelpers          $helpers,
+        ContentInterface          $content,
+        EmbedParser               $embedParser,
+        HtmlArticleLocker         $articleLocker,
+        EngineFactory             $engineFactory
     ) {
         $this->mailSourceTemplateRepository = $mailSourceTemplateRepository;
         $this->helpers = $helpers;
@@ -120,7 +120,7 @@ class MediaBriefingGenerator implements IGenerator
         $text = str_replace("<br />", "\r\n", $post);
         $lockedText = str_replace("<br />", "\r\n", $lockedPost);
 
-        $lockedPost = $this->articleLocker->putLockedMessage($lockedPost);
+        $lockedPost = $this->articleLocker->injectLockedMessage($lockedPost);
 
         $text = strip_tags($text);
         $lockedText = strip_tags($lockedText);