+++ Mailer/app/Models/Generators/NapunkNewsfilterGenerator.php
@@ -2,13 +2,6 @@
 
 namespace Remp\Mailer\Models\Generators;
 
-use Remp\MailerModule\Models\ContentGenerator\Engine\EngineFactory;
-use Remp\MailerModule\Models\Generators\ArticleLocker;
-use Remp\MailerModule\Models\Generators\EmbedParser;
-use Remp\MailerModule\Models\Generators\WordpressHelpers;
-use Remp\MailerModule\Models\PageMeta\Content\ContentInterface;
-use Remp\MailerModule\Repositories\SourceTemplatesRepository;
-
 class NapunkNewsfilterGenerator extends NewsfilterGenerator
 {
     public function process(array $values): array